+++ src/shared/forms/user.js
@@ -1,13 +1,18 @@
-import Form from 'mobx-ajv-form';
-import schema from '~/src/shared/schemas/user';
+import Form from 'mobx-react-form';
+import validatorjs from 'validatorjs';
 import { dispatch } from '~/src/utils/state';
 
 class UserForm extends Form {
 
   handleOnSubmit = (e) => {
     e.preventDefault();
-    if (!this.validate()) return;
 
+    this.validate()
+      .then((isValid) =>
+        isValid && this.onSuccess());
+  }
+
+  onSuccess() {
     dispatch('auth.register', this.values())
       .then(() => dispatch('ui.authModal.toggleSection', 'signin'))
       .then(() => dispatch('ui.snackBar.open', 'Register Successful.'))
@@ -18,16 +23,21 @@ class UserForm extends Form {
 
 export default
   new UserForm({
-    schema,
+    plugins: {
+      dvr: validatorjs,
+    },
     fields: {
       username: {
         label: 'Username',
+        rules: 'required|string|between:5,20',
       },
       email: {
         label: 'Email',
+        rules: 'required|email|string|between:5,20',
       },
       password: {
         label: 'Password',
+        rules: 'required|string|between:5,20',
       },
     },
   });