+++ src/shared/stores.js
@@ -1,5 +1,5 @@
 import { store } from '~/src/utils/state';
-import { useStrict } from 'mobx';
+import { isStrictModeEnabled } from 'mobx';
 
 import UIStore from './stores/ui';
 import AppStore from './stores/app';
@@ -7,11 +7,11 @@ import AuthStore from './stores/auth';
 import PostStore from './stores/post';
 
 /**
-  Enables / disables strict mode globally.
+  Enables MobX strict mode globally.
   In strict mode, it is not allowed to
   change any state outside of an action
  */
-useStrict(true);
+isStrictModeEnabled();
 
 /**
   Stores