+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 # 0.6.0 (alpha.6)
 
-* Form management with mobx-ajv-form
+* Form management with mobx-react-form
 * Updated Graphic
 
 # 0.5.0 (alpha.5)